requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-api-python-client
google-auth
google-cloud-aiplatform
